Move to net.imglib2.fluent package. Start systematically implementing Views methods
