+++ src/main/java/net/imglib2/view/fluent/RaiView.java
@@ -1,9 +1,11 @@
-package net.imglib2.streamifiedview;
+package net.imglib2.view.fluent;
 
 import net.imglib2.Cursor;
+import net.imglib2.FlatIterationOrder;
 import net.imglib2.Interval;
 import net.imglib2.RandomAccess;
 import net.imglib2.RandomAccessibleInterval;
+import net.imglib2.view.IterableRandomAccessibleInterval;
 import net.imglib2.view.Views;
 
 /**
@@ -21,8 +23,8 @@ import net.imglib2.view.Views;
  * its {@link #delegate} and return {@code RaiView}, {@code RaView}, or {@code
  * RraView} wrappers.
  * <p>
- * This provides a fluent interface for conveniently chaining {@code Views}
- * methods. For example
+ * This provides a fluent API for conveniently chaining {@code Views} methods.
+ * For example
  * <pre>
  * {@code RandomAccessibleInterval< IntType > view =
  *                img.view()
@@ -38,6 +40,77 @@ import net.imglib2.view.Views;
  */
 public interface RaiView< T > extends RaView< T, RaiView< T > >, RandomAccessibleInterval< T >
 {
+	@Override
+	RandomAccessibleInterval< T > delegate();
+
+	static < T > RaiView< T > wrap( final RandomAccessibleInterval< T > delegate )
+	{
+		return new RaiWrapper<>( delegate );
+	}
+
+	// -- Views methods -------------------------------------------------------
+
+	/**
+	 * Enforce {@link FlatIterationOrder} order for this {@code
+	 * RandomAccessibleInterval}.
+	 * <p>
+	 * If this passed {@code RandomAccessibleInterval} already has flat
+	 * iteration order then it is returned directly. If not, then it is wrapped
+	 * in a {@link IterableRandomAccessibleInterval}.
+	 *
+	 * @return a view with flat iteration order
+	 */
+	default RaiView< T > flatIterable()
+	{
+		return wrap( Views.flatIterable( delegate() ) );
+	}
+
+	/**
+	 * Take a <em>(n-1)</em>-dimensional slice of this <em>n</em>-dimensional
+	 * {@code RandomAccessible}, by fixing the {@code d} dimension of
+	 * coordinates to {@code pos}.
+	 *
+	 * @param d
+	 * 		coordinate dimension to fix
+	 * @param pos
+	 * 		coordinate value to fix {@code d}th dimension to
+	 *
+	 * @return a view on the given slice
+	 */
+	@Override
+	default RaiView< T > slice( int d, long pos )
+	{
+		return wrap( Views.hyperSlice( delegate(), d, pos ) );
+	}
+
+	/**
+	 * Create a <em>(n+1)</em>-dimensional view of this <em>n</em>-dimensional
+	 * {@code RandomAccessibleInterval}, by replicating values along the added
+	 * axis.
+	 * <p>
+	 * The additional dimension is the last dimension. For example, an XYZ view
+	 * is created for an XY source. When accessing an XYZ sample in the view,
+	 * the final coordinate is discarded and the source XY sample is accessed.
+	 *
+	 * @param minOfNewDim
+	 * 		interval min in the added dimension.
+	 * @param maxOfNewDim
+	 * 		interval max in the added dimension.
+	 *
+	 * @return a view with an additional dimension
+	 */
+	default RaiView< T > addDimension( long minOfNewDim, long maxOfNewDim )
+	{
+		return wrap( Views.addDimension( delegate(), minOfNewDim, maxOfNewDim ) );
+	}
+
+
+
+	// done until here
+	//////////////////
+
+
+
 	default RaiView< T > expandValue( final T value, long... border )
 	{
 		return wrap( Views.expandValue( delegate(), value, border ) );
@@ -63,23 +136,18 @@ public interface RaiView< T > extends RaView< T, RaiView< T > >, RandomAccessibl
 		return wrap( Views.rotate( delegate(), fromAxis, toAxis ) );
 	}
 
-	default RaiView< T > hyperSlice( int d, long pos )
-	{
-		return wrap( Views.hyperSlice( delegate(), d, pos ) );
-	}
-
 	default RaView< T, ? > extendBorder()
 	{
 		return RaView.wrap( Views.extendBorder( delegate() ) );
 	}
 
-	@Override
-	RandomAccessibleInterval< T > delegate();
 
-	static < T > RaiView< T > wrap( final RandomAccessibleInterval< T > delegate )
-	{
-		return new RaiWrapper<>( delegate );
-	}
+
+
+
+
+
+	// -- RandomAccessible ----------------------------------------------------
 
 	@Override
 	default RaiView< T > view()
