+++ src/test/java/net/imglib2/view/fluent/StreamifiedViewsExample.java
@@ -1,4 +1,4 @@
-package net.imglib2.streamifiedview;
+package net.imglib2.view.fluent;
 
 import net.imglib2.Cursor;
 import net.imglib2.RandomAccess;
